+++ docs/src/doc/docs/whats-new.md
@@ -1,8 +1,14 @@
-# What's New in Docling Java {{ gradle.project_version }}
+# What's New in Docling Java
 
 Docling Java {{ gradle.project_version }} provides a number of new features, enhancements, and bug fixes. This page includes the highlights of the release, but you can also check out the full [release notes](https://github.com/docling-project/docling-java/releases) for more details about each new feature and bug fix.
 
 ## Docling Serve
+
+### {{ gradle.project_version }}
+
+* Implementation of the Docling Serve clear and task APIs in `docling-serve-api` and `docling-serve-client`.
+
+### 0.3.0
 
 * The classes in the `docling-serve-api` module have been moved from the `ai.docling.api.serve` package to the `ai.docling.serve.api` package.
 * The classes in the `docling-serve-client` module have been moved from the `ai.docling.client.serve` package to the `ai.docling.serve.client` package.